Change tests so that they pass. todo: make them valid too
